backend_django/requirements.txt
# Core Django
# Target Python runtime: 3.11.x (see ../runtime.txt)
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
whitenoise==6.6.0
django-filter==23.5
drf-spectacular==0.27.1

# Database
psycopg2-binary==2.9.9
django-extensions==3.2.3

# WebSockets & Async
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Authentication & Security
djangorestframework-simplejwt==5.3.1
PyJWT==2.10.1
bcrypt==4.1.2
cryptography==42.0.5

# Third-party integrations
cloudinary==1.44.1
django-cloudinary-storage==0.3.0
qrcode==7.4.2
Pillow==10.3.0
requests==2.31.0
urllib3<2  # Keep urllib3 compatible with LibreSSL-based Python runtimes
django-anymail[sendgrid]==9.2

# Caching & Sessions
django-redis==5.4.0
redis==5.0.1
celery==5.6.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# File handling
openpyxl==3.1.5
pandas==2.1.3

# Monitoring & Logging
python-json-logger==2.0.7
sentry-sdk==1.39.1

# Deployment
gunicorn==21.2.0
uvicorn[standard]==0.29.0   # Required for UvicornWorker (ASGI/WebSocket support)
dj-database-url==2.1.0

# === DEV-ONLY DEPENDENCIES ===
# Install these locally with: pip install -r requirements-dev.txt
# Do NOT install in production to save disk and build time.
# django-debug-toolbar==4.2.0
# django-silk==5.0.4
# faker==20.1.0
# black==23.12.1
# flake8==6.1.0
# isort==5.13.2
# pylint==3.0.3
# pyre-check==0.9.25

pywebpush==1.14.0
xhtml2pdf==0.2.11
httpx==0.27.0
